requirements.txt
pymupdf4llm>=0.0.1
pymupdf>=1.22.0
websockets>=11.0
httpx>=0.24.0
Pillow>=10.0.0
pyautogui>=0.9.54
psutil>=5.9.0
GPUtil>=1.4.0
pyperclip>=1.8.2
flask>=3.0.0
requests>=2.31.0
discord.py>=2.3.0
python-telegram-bot>=20.0
cryptography>=41.0.0
imageio-ffmpeg>=0.4.9
python-dotenv>=1.0.0
aiohttp>=3.9.0
uvicorn>=0.22.0
pylatex>=1.4.1
numpy>=1.24.0
soundfile>=0.12.1
pydantic>=2.0.0
openai>=1.0.0
anthropic>=0.5.0
python-multipart>=0.0.6
pocket-tts>=1.0.3
